@@ -1,7 +1,7 @@
 import 'package:r_flutter/src/generator/generator.dart';
 import 'package:r_flutter/src/model/dart_class.dart';
 
-DartClass generateFontClass(List<String> fonts) {
+DartClass? generateFontClass(List<String> fonts) {
   if (fonts.isEmpty) {
     return null;
   }
